+++ PDV/modules/caja/manifest.js
@@ -3,8 +3,8 @@
  * PDV
  * Archivo: manifest.js
  * Módulo: caja
- * Descripción: Administración básica de Caja.
- * Versión: 0.9.5
+ * Descripción: Administración de Caja e historial de sesiones.
+ * Versión: 0.9.13
  * ==========================================================
  */
 
@@ -25,7 +25,7 @@ export default {
 
     enabled: true,
 
-    version: "0.9.3",
+    version: "0.9.13",
 
     author: "Ricardo Farabundo Magallón Juan-Qui"